+++ src/styles/default.ts
@@ -19,6 +19,7 @@ const style = registerStyle(
         init: "INIT",
         crash: "CRASH",
         group: "GROUP",
+        timer: "TIMER",
 
         info_secondary: "INFO",
         warning_secondary: "WARNING",
@@ -38,6 +39,7 @@ const style = registerStyle(
             "init": consoleColors.FgWhite,
             "crash": consoleColors.FgRed,
             "group": consoleColors.FgGray,
+            "timer": consoleColors.FgGray,
 
             "info_secondary": consoleColors.FgWhite,
             "warning_secondary": consoleColors.FgYellow,
@@ -47,6 +49,7 @@ const style = registerStyle(
             "init_secondary": consoleColors.FgWhite,
             "crash_secondary": consoleColors.FgRed,
             "group_secondary": consoleColors.FgGray,
+            "timer_secondary": consoleColors.FgGray,
 
             "date": consoleColors.FgGray,
             "who": consoleColors.FgMagenta,